# Conflicts:
#	contracts/gas-snapshots/workflow.gas-snapshot
#	gethwrappers/workflow/generated/capabilities_registry_wrapper_v2_dev/capabilities_registry_wrapper_v2_dev.go
#	gethwrappers/workflow/generation/generated-wrapper-dependency-versions-do-not-edit.txt

+++ contracts/src/v0.8/automation/testhelpers/KeeperConsumer.sol
@@ -1,14 +1,16 @@
 pragma solidity 0.8.16;
 
-import "../interfaces/KeeperCompatibleInterface.sol";
 import "../KeeperBase.sol";
+import "../interfaces/KeeperCompatibleInterface.sol";
 
 contract KeeperConsumer is KeeperCompatibleInterface, KeeperBase {
-  uint public counter;
-  uint public immutable interval;
-  uint public lastTimeStamp;
+  uint256 public counter;
+  uint256 public immutable interval;
+  uint256 public lastTimeStamp;
 
-  constructor(uint updateInterval) public {
+  constructor(
+    uint256 updateInterval
+  ) public {
     interval = updateInterval;
     lastTimeStamp = block.timestamp;
     counter = 0;
@@ -20,7 +22,9 @@ contract KeeperConsumer is KeeperCompatibleInterface, KeeperBase {
     return (true, checkData);
   }
 
-  function performUpkeep(bytes calldata performData) external override {
+  function performUpkeep(
+    bytes calldata performData
+  ) external override {
     counter = counter + 1;
   }
 }
